ai_engine/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
supabase
python-dotenv
docling
pydantic
openai 
yandex-chain 
tenacity # Для ретраев